+++ controltest/functions.h
@@ -1,2 +1,2 @@
 extern void set_flag(bool in_flag);
-extern void get_data(double *req_data, unsigned int *data_size);
+extern void get_data(float *req_data, unsigned int *data_size);
